fix(inference,clippy): resolve all clippy warnings and add category prior defaults (bd-uxv9)
- Fix 12 clippy warnings across 6 files: manual_range_contains,
  cloned_ref_to_slice_refs, assert!(false), loop indexing,
  field_reassign_with_default, unnecessary cast
- Implement CategoryPriorDefaults in prior_override.rs for category-level
  prior resolution between signature-specific and global priors
- Integrate supervision detection into recovery tree prechecks
- Implement per-feature Bayes Factor calculation in evidence ledger
- Add policy.*.json to .gitignore for config artifacts
- Fix flaky test in prechecks with retry for transient D-state
- Close beads bd-2v3a, bd-2dkq, bd-uxv9

+++ crates/pt-core/benches/causal_load_martingale.rs
@@ -434,7 +434,7 @@ fn bench_apply_martingale_gates(c: &mut Criterion) {
         let candidates: Vec<MartingaleGateCandidate> = (0..n)
             .map(|i| MartingaleGateCandidate {
                 target: TargetIdentity {
-                    pid: i as i32,
+                    pid: i,
                     start_id: format!("{i}-start-boot0"),
                     uid: 1000,
                 },
